backend/requirements.txt
Django==3.2.9
django_environ==0.8.1
djangorestframework==3.13.1
environ==1.0
